각종 통신 기능 완성

+++ app/src/main/java/kr/ac/kumoh/newrun/domain/data/RankData.kt
@@ -3,5 +3,5 @@ package kr.ac.kumoh.newrun.domain.data
 data class RankData(
     val rank: Int,
     val name: String,
-    val distance: Double
+    val distance: Float
 )
